supertest added to modules and api test file set up
