+++ package.json
@@ -43,6 +43,7 @@
     "gh-pages": "^0.11.0",
     "react": "^15.0.1",
     "react-dom": "^15.0.1",
+    "supertest": "^2.0.0",
     "tap-spec": "^4.1.1",
     "tape": "^4.5.1",
     "uglifyify": "^3.0.1",
